backend/requirements.txt
fastapi
uvicorn[standard]
pipecat-ai[fish,deepgram,openai]
python-dotenv
fish-audio-sdk
python-multipart
requests
websockets
openai-whisper
pyannote.audio
torch
torchaudio
numpy
scipy

